test play,pause,progress bar update
