+++ index.html
@@ -128,6 +128,6 @@
                     </div>
                </section>
             </main>
-        <script src="scripts.js"></script>
+        <script type="module" src="scripts.js"></script>
     </body>
 </html>
